Camera based turn and fixing all animations
Way more smooth now
