--- SolarFlare/Assets/Models/Katana/Mesh.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b3fb945ff983abd4b97041af410fc992
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
