+++ SolarFlare/Assets/Scripts/tellManagerAboutMe.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19a9c358eb157784794e342e11d21d98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: -200
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
